Reorganize e2e test utils
