+++ e2e/agent/agent_e2e_test.go
@@ -11,6 +11,8 @@ import (
 
 	"github.com/grafana/xk6-disruptor/pkg/kubernetes"
 	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/checks"
+	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/cluster"
+	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/deploy"
 	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/fixtures"
 	"github.com/grafana/xk6-disruptor/pkg/testutils/kubernetes/builders"
 
@@ -103,10 +105,10 @@ func buildGrpcbinPodWithDisruptorAgent(cmd []string) *corev1.Pod {
 }
 
 func Test_Agent(t *testing.T) {
-	cluster, err := fixtures.BuildE2eCluster(
-		fixtures.DefaultE2eClusterConfig(),
-		fixtures.WithName("e2e-xk6-agent"),
-		fixtures.WithIngressPort(30080),
+	cluster, err := cluster.BuildE2eCluster(
+		cluster.DefaultE2eClusterConfig(),
+		cluster.WithName("e2e-xk6-agent"),
+		cluster.WithIngressPort(30080),
 	)
 	if err != nil {
 		t.Errorf("failed to create e2e cluster: %v", err)
@@ -171,7 +173,7 @@ func Test_Agent(t *testing.T) {
 				}
 				defer k8s.Client().CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
 
-				err = fixtures.DeployApp(
+				err = deploy.ExposeApp(
 					k8s,
 					namespace,
 					tc.pod,
@@ -203,7 +205,7 @@ func Test_Agent(t *testing.T) {
 		}
 		defer k8s.Client().CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
 
-		err = fixtures.RunPod(
+		err = deploy.RunPod(
 			k8s,
 			namespace,
 			buildHttpbinPodWithDisruptorAgent(injectHTTP500),
