+++ e2e/kubernetes/kubernetes_e2e_test.go
@@ -11,7 +11,9 @@ import (
 
 	"github.com/grafana/xk6-disruptor/pkg/kubernetes"
 	"github.com/grafana/xk6-disruptor/pkg/kubernetes/helpers"
-	"github.com/grafana/xk6-disruptor/pkg/testutils/cluster"
+	kindcluster "github.com/grafana/xk6-disruptor/pkg/testutils/cluster"
+	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/cluster"
+	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/deploy"
 	"github.com/grafana/xk6-disruptor/pkg/testutils/e2e/fixtures"
 
 	corev1 "k8s.io/api/core/v1"
@@ -19,10 +21,10 @@ import (
 )
 
 func Test_Kubernetes(t *testing.T) {
-	cluster, err := fixtures.BuildE2eCluster(
-		fixtures.DefaultE2eClusterConfig(),
-		fixtures.WithName("e2e-kubernetes"),
-		fixtures.WithIngressPort(30081),
+	cluster, err := cluster.BuildE2eCluster(
+		cluster.DefaultE2eClusterConfig(),
+		cluster.WithName("e2e-kubernetes"),
+		cluster.WithIngressPort(30081),
 	)
 	if err != nil {
 		t.Errorf("failed to create cluster: %v", err)
@@ -107,7 +109,7 @@ func Test_Kubernetes(t *testing.T) {
 		}
 		defer k8s.Client().CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
 
-		err = fixtures.RunPod(k8s, namespace, fixtures.BuildBusyBoxPod(), 10*time.Second)
+		err = deploy.RunPod(k8s, namespace, fixtures.BuildBusyBoxPod(), 10*time.Second)
 		if err != nil {
 			t.Errorf("error creating pod: %v", err)
 			return
@@ -139,7 +141,7 @@ func Test_Kubernetes(t *testing.T) {
 		}
 		defer k8s.Client().CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
 
-		err = fixtures.RunPod(k8s, namespace, fixtures.BuildPausedPod(), 10*time.Second)
+		err = deploy.RunPod(k8s, namespace, fixtures.BuildPausedPod(), 10*time.Second)
 		if err != nil {
 			t.Errorf("error running pod %v: ", err)
 			return
@@ -189,9 +191,10 @@ func Test_Kubernetes(t *testing.T) {
 }
 
 func Test_UnsupportedKubernetesVersion(t *testing.T) {
-	config, err := cluster.NewConfig(
+	// TODO: use e2e cluster. This will require an option for setting the K8s version in the e2e cluster
+	config, err := kindcluster.NewConfig(
 		"e2e-v1-22-0-cluster",
-		cluster.Options{
+		kindcluster.Options{
 			Version: "v1.22.0",
 			Wait:    time.Second * 60,
 		},
