--- pkg/testutils/e2e/fixtures/docs.go
@@ -1,2 +1,0 @@
-// Package fixtures implements helpers for setting e2e tests
-package fixtures
