+++ pkg/testutils/e2e/fixtures/fixtures.go
@@ -1,15 +1,10 @@
+// Package fixtures implements helpers for setting e2e tests
 package fixtures
 
 import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/grafana/xk6-disruptor/pkg/kubernetes"
 	"github.com/grafana/xk6-disruptor/pkg/testutils/kubernetes/builders"
 
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
@@ -156,96 +151,4 @@ func BuildNginxService() *corev1.Service {
 			},
 		).
 		Build()
-}
-
-// ExposeService creates a service and waits for it to be ready before exposing as an ingress.
-// The ingress routes request that specify the service's name as host to this service.
-func ExposeService(
-	k8s kubernetes.Kubernetes,
-	namespace string,
-	svc *corev1.Service,
-	port intstr.IntOrString,
-	timeout time.Duration,
-) error {
-	_, err := k8s.Client().CoreV1().Services(namespace).Create(
-		context.TODO(),
-		svc,
-		metav1.CreateOptions{},
-	)
-	if err != nil {
-		return fmt.Errorf("failed to create service %s: %w", svc.Name, err)
-	}
-
-	// wait for the service to be ready for accepting requests
-	err = k8s.ServiceHelper(namespace).WaitServiceReady(context.TODO(), svc.Name, timeout)
-	if err != nil {
-		return fmt.Errorf("error waiting for service %s: %w", svc.Name, err)
-	}
-
-	ingress := builders.NewIngressBuilder(svc.Name, port).
-		WithNamespace(namespace).
-		WithHost(svc.Name).
-		Build()
-
-	_, err = k8s.Client().NetworkingV1().Ingresses(namespace).Create(
-		context.TODO(),
-		ingress,
-		metav1.CreateOptions{},
-	)
-	if err != nil {
-		return fmt.Errorf("failed to create ingress %s: %w", ingress.Name, err)
-	}
-
-	// wait for the ingress to be ready for accepting requests
-	err = k8s.ServiceHelper(namespace).WaitIngressReady(context.TODO(), svc.Name, timeout)
-	if err != nil {
-		return fmt.Errorf("error waiting for ingress %s: %w", ingress.Name, err)
-	}
-
-	return nil
-}
-
-// RunPod creates a pod and waits it for be running
-func RunPod(k8s kubernetes.Kubernetes, ns string, pod *corev1.Pod, timeout time.Duration) error {
-	_, err := k8s.Client().CoreV1().Pods(ns).Create(
-		context.TODO(),
-		pod,
-		metav1.CreateOptions{},
-	)
-	if err != nil {
-		return fmt.Errorf("error creating pod %s: %w", pod.Name, err)
-	}
-
-	running, err := k8s.PodHelper(ns).WaitPodRunning(
-		context.TODO(),
-		pod.Name,
-		timeout,
-	)
-	if err != nil {
-		return fmt.Errorf("error waiting for pod %s: %w", pod.Name, err)
-	}
-	if !running {
-		return fmt.Errorf("pod %s not ready after %f: ", pod.Name, timeout.Seconds())
-	}
-
-	return nil
-}
-
-// DeployApp deploys a pod in a namespace and exposes it as a service in a cluster
-func DeployApp(
-	k8s kubernetes.Kubernetes,
-	ns string,
-	pod *corev1.Pod,
-	svc *corev1.Service,
-	port intstr.IntOrString,
-	timeout time.Duration,
-) error {
-	start := time.Now()
-	err := RunPod(k8s, ns, pod, timeout)
-	if err != nil {
-		return fmt.Errorf("failed to create pod %s in namespace %s: %w", pod.Name, ns, err)
-	}
-
-	timeLeft := timeout - time.Since(start)
-	return ExposeService(k8s, ns, svc, port, timeLeft)
 }
